clean up some lazy typing

+++ src/de.rs
@@ -83,7 +83,7 @@ pub fn decode(adc: &str) -> Result<DeserializedDeck, String> {
 ///}
 ///```
 ///
-pub fn map_card_ids_to_cards(sets: Vec<&str>) -> Result<HashMap<usize, crate::Card>, String> {
+pub fn map_card_ids_to_cards(sets: Vec<&str>) -> Result<HashMap<u32, crate::Card>, String> {
     let mut d_sets = Vec::new();
     for set in sets {
         let s: crate::CardSetJson = match serde_json::from_str(set) {
@@ -100,8 +100,8 @@ pub fn map_card_ids_to_cards(sets: Vec<&str>) -> Result<HashMap<usize, crate::Ca
 
     Ok(set_up_deck_map(d_sets))
 }
-fn set_up_deck_map(sets: Vec<crate::CardSet>) -> HashMap<usize, crate::Card> {
-    let mut map = HashMap::<usize, crate::Card>::new();
+fn set_up_deck_map(sets: Vec<crate::CardSet>) -> HashMap<u32, crate::Card> {
+    let mut map = HashMap::<u32, crate::Card>::new();
     for set in sets {
         for card in set.card_list {
             map.insert(card.card_id, card);
@@ -111,13 +111,13 @@ fn set_up_deck_map(sets: Vec<crate::CardSet>) -> HashMap<usize, crate::Card> {
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct DeserializedHero {
-    pub id: usize,
-    pub turn: usize,
+    pub id: u32,
+    pub turn: u32,
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct DeserializedCard {
-    pub id: usize,
-    pub count: usize,
+    pub id: u32,
+    pub count: u32,
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct DeserializedDeck {
@@ -146,23 +146,23 @@ fn parse_deck(_deck_code: String, deck_bytes: Vec<u8>) -> Result<DeserializedDec
 
     let mut num_heroes = 0;
     read_encoded_u32(
-        *version_and_heroes as usize,
+        *version_and_heroes as u32,
         3,
         &deck_bytes,
         &mut current_byte_index,
-        total_card_bytes as usize,
+        total_card_bytes,
         &mut num_heroes,
     );
 
     let mut heroes = Vec::<DeserializedHero>::new();
     let mut prev_card_base = 0;
     for _curr_hero in 0..num_heroes {
-        let mut hero_turn = 0 as usize;
-        let mut hero_card_id = 0 as usize;
+        let mut hero_turn = 0 as u32;
+        let mut hero_card_id = 0 as u32;
         if !read_serialized_card(
             &deck_bytes,
             &mut current_byte_index,
-            total_card_bytes as usize,
+            total_card_bytes,
             &mut prev_card_base,
             &mut hero_turn,
             &mut hero_card_id,
@@ -179,13 +179,13 @@ fn parse_deck(_deck_code: String, deck_bytes: Vec<u8>) -> Result<DeserializedDec
 
     let mut cards = Vec::<DeserializedCard>::new();
     prev_card_base = 0;
-    while current_byte_index < total_card_bytes as usize {
+    while current_byte_index < total_card_bytes {
         let mut card_count = 0;
         let mut card_id = 0;
         if !read_serialized_card(
             &deck_bytes,
             &mut current_byte_index,
-            total_card_bytes as usize,
+            total_card_bytes,
             &mut prev_card_base,
             &mut card_count,
             &mut card_id,
@@ -217,26 +217,21 @@ fn parse_deck(_deck_code: String, deck_bytes: Vec<u8>) -> Result<DeserializedDec
     })
 }
 
-fn read_bits_chunk(
-    n_chunk: usize,
-    n_bits: usize,
-    n_curr_shift: usize,
-    n_out_bits: &mut usize,
-) -> bool {
+fn read_bits_chunk(n_chunk: u32, n_bits: u32, n_curr_shift: u32, n_out_bits: &mut u32) -> bool {
     let continue_bit = 1 << n_bits;
     let new_bits = n_chunk & (continue_bit - 1);
-    *n_out_bits |= (new_bits << n_curr_shift) as usize;
+    *n_out_bits |= (new_bits << n_curr_shift) as u32;
 
     n_chunk & continue_bit != 0
 }
 
 fn read_encoded_u32(
-    base_value: usize,
-    base_bits: usize,
+    base_value: u32,
+    base_bits: u32,
     deck_bytes: &Vec<u8>,
     start_index: &mut usize,
     end_index: usize,
-    out_value: &mut usize,
+    out_value: &mut u32,
 ) {
     *out_value = 0;
     let mut delta_shift = 0;
@@ -250,7 +245,7 @@ fn read_encoded_u32(
 
             let next_byte = deck_bytes.get(*start_index).unwrap();
             *start_index += 1;
-            if !read_bits_chunk(*next_byte as usize, 7, delta_shift, out_value) {
+            if !read_bits_chunk(*next_byte as u32, 7, delta_shift, out_value) {
                 break;
             }
 
@@ -263,9 +258,9 @@ fn read_serialized_card(
     deck_bytes: &Vec<u8>,
     start_index: &mut usize,
     end_index: usize,
-    prev_card_base: &mut usize,
-    out_count: &mut usize,
-    out_id: &mut usize,
+    prev_card_base: &mut u32,
+    out_count: &mut u32,
+    out_id: &mut u32,
 ) -> bool {
     //end of the memory block?
     if *start_index > end_index {
@@ -283,7 +278,7 @@ fn read_serialized_card(
     //read in the delta, which has 5 bits in the header, then additional bytes while the value is set
     let mut card_delta = 0;
     read_encoded_u32(
-        *header as usize,
+        *header as u32,
         5,
         &deck_bytes,
         start_index,
@@ -300,7 +295,7 @@ fn read_serialized_card(
         }
         false => {
             //the count is just the upper two bits + 1 (since we don't encode zero)
-            *out_count = (header >> 6) as usize + 1;
+            *out_count = (header >> 6) as u32 + 1;
         }
     }
